# src/main/resources/openapi.yaml
openapi: 3.0.3
info:
  title: Backend Bank API
  version: 1.0.0
  description: API REST para gestion de clientes, cuentas, movimientos y reportes.
servers:
  - url: http://localhost:8080
    description: Local
  - url: http://{host}:{port}
    description: Entorno configurable
    variables:
      host:
        default: localhost
      port:
        default: "8080"
security:
  - basicAuth: []
tags:
  - name: Clientes
  - name: Cuentas
  - name: Movimientos
  - name: Reportes

paths:
  /clientes:
    get:
      tags: [Clientes]
      summary: Listar clientes
      operationId: findAllClients
      responses:
        "200":
          description: Lista de clientes
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ClientResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Clientes]
      summary: Crear cliente
      operationId: createClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientCreateRequest"
      responses:
        "201":
          description: Cliente creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "409":
          $ref: "#/components/responses/Conflict"

  /clientes/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Clientes]
      summary: Obtener cliente por id
      operationId: findClientById
      responses:
        "200":
          description: Cliente encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Clientes]
      summary: Reemplazar cliente
      operationId: updateClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientUpdateRequest"
      responses:
        "200":
          description: Cliente actualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    patch:
      tags: [Clientes]
      summary: Actualizar parcialmente cliente
      operationId: patchClient
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClientPatchRequest"
      responses:
        "200":
          description: Cliente actualizado parcialmente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Clientes]
      summary: Eliminar cliente
      operationId: deleteClient
      responses:
        "204":
          description: Cliente eliminado
        "404":
          $ref: "#/components/responses/NotFound"

  /cuentas:
    get:
      tags: [Cuentas]
      summary: Listar cuentas
      operationId: findAllAccounts
      responses:
        "200":
          description: Lista de cuentas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AccountResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Cuentas]
      summary: Crear cuenta
      operationId: createAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountCreateRequest"
      responses:
        "201":
          description: Cuenta creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"

  /cuentas/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Cuentas]
      summary: Obtener cuenta por id
      operationId: findAccountById
      responses:
        "200":
          description: Cuenta encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Cuentas]
      summary: Reemplazar cuenta
      operationId: updateAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountUpdateRequest"
      responses:
        "200":
          description: Cuenta actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    patch:
      tags: [Cuentas]
      summary: Actualizar parcialmente cuenta
      operationId: patchAccount
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AccountPatchRequest"
      responses:
        "200":
          description: Cuenta actualizada parcialmente
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AccountResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
    delete:
      tags: [Cuentas]
      summary: Eliminar cuenta
      operationId: deleteAccount
      responses:
        "204":
          description: Cuenta eliminada
        "404":
          $ref: "#/components/responses/NotFound"

  /movimientos:
    get:
      tags: [Movimientos]
      summary: Listar movimientos
      operationId: findAllTransactions
      responses:
        "200":
          description: Lista de movimientos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransactionResponse"
        "500":
          $ref: "#/components/responses/InternalError"
    post:
      tags: [Movimientos]
      summary: Registrar movimiento
      operationId: createTransaction
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TransactionCreateRequest"
      responses:
        "201":
          description: Movimiento creado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
        "422":
          $ref: "#/components/responses/BusinessRuleError"

  /movimientos/{id}:
    parameters:
      - $ref: "#/components/parameters/IdPath"
    get:
      tags: [Movimientos]
      summary: Obtener movimiento por id
      operationId: findTransactionById
      responses:
        "200":
          description: Movimiento encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Movimientos]
      summary: Eliminar movimiento
      operationId: deleteTransaction
      responses:
        "204":
          description: Movimiento eliminado
        "404":
          $ref: "#/components/responses/NotFound"

  /reportes:
    get:
      tags: [Reportes]
      summary: Generar reporte por cliente y rango de fechas
      operationId: generateReport
      parameters:
        - name: clienteId
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: fechaDesde
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: fechaHasta
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: formato
          in: query
          required: false
          schema:
            type: string
            enum: [json, pdf]
            default: json
      responses:
        "200":
          description: Reporte generado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReporteResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"

components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic

  parameters:
    IdPath:
      name: id
      in: path
      required: true
      schema:
        type: integer
        format: int64

  responses:
    BadRequest:
      description: Solicitud invalida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorModel"
    NotFound:
      description: Recurso no encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorModel"
    Conflict:
      description: Conflicto de negocio o datos duplicados
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorModel"
    BusinessRuleError:
      description: Regla de negocio no cumplida
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorModel"
    InternalError:
      description: Error interno no controlado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorModel"

  schemas:
    ErrorModel:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-04-25T12:45:08Z"
        status:
          type: integer
          example: 422
        code:
          type: string
          example: "BBB0005"
        title:
          type: string
          example: "INSUFFICIENT_FUNDS"
        instance:
          type: string
          example: "BBB0005"
        message:
          type: string
          example: "Saldo no disponible"
        detail:
          type: string
          example: "Saldo no disponible"
        type:
          type: string
          example: "Unprocessable Content"
        path:
          type: string
          example: "/movimientos"
        traceId:
          type: string
          example: "b1f09f6d-6f55-4be8-a7b4-44c13b0cfc77"
        fieldErrors:
          type: array
          items:
            $ref: "#/components/schemas/FieldErrorItem"
      required: [timestamp, status, code, title, message, path]

    FieldErrorItem:
      type: object
      properties:
        field:
          type: string
          example: "numeroCuenta"
        message:
          type: string
          example: "must not be blank"
      required: [field, message]

    Gender:
      type: string
      enum: [MASCULINO, FEMENINO, OTRO]

    AccountType:
      type: string
      enum: [AHORROS, CORRIENTE]

    TransactionType:
      type: string
      enum: [CREDIT, DEBIT]

    ClientCreateRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 120
        genero:
          $ref: "#/components/schemas/Gender"
        edad:
          type: integer
          minimum: 0
          maximum: 130
        identificacion:
          type: string
          maxLength: 50
        direccion:
          type: string
          maxLength: 180
        telefono:
          type: string
          maxLength: 30
        contrasena:
          type: string
          minLength: 8
          maxLength: 255
        estado:
          type: boolean
          default: true
      required: [nombre, genero, edad, identificacion, direccion, telefono, contrasena]

    ClientUpdateRequest:
      $ref: "#/components/schemas/ClientCreateRequest"

    ClientPatchRequest:
      type: object
      properties:
        nombre:
          type: string
          maxLength: 120
        genero:
          $ref: "#/components/schemas/Gender"
        edad:
          type: integer
          minimum: 0
          maximum: 130
        identificacion:
          type: string
          maxLength: 50
        direccion:
          type: string
          maxLength: 180
        telefono:
          type: string
          maxLength: 30
        contrasena:
          type: string
          minLength: 8
          maxLength: 255
        estado:
          type: boolean

    ClientResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        nombre:
          type: string
        genero:
          $ref: "#/components/schemas/Gender"
        edad:
          type: integer
        identificacion:
          type: string
        direccion:
          type: string
        telefono:
          type: string
        estado:
          type: boolean
      required: [id, nombre, genero, edad, identificacion, direccion, telefono, estado]

    AccountCreateRequest:
      type: object
      properties:
        clienteId:
          type: integer
          format: int64
        numeroCuenta:
          type: string
          maxLength: 30
        tipoCuenta:
          $ref: "#/components/schemas/AccountType"
        saldoInicial:
          type: number
          format: double
          minimum: 0
        estado:
          type: boolean
          default: true
      required: [clienteId, numeroCuenta, tipoCuenta, saldoInicial]

    AccountUpdateRequest:
      $ref: "#/components/schemas/AccountCreateRequest"

    AccountPatchRequest:
      type: object
      properties:
        numeroCuenta:
          type: string
          maxLength: 30
        tipoCuenta:
          $ref: "#/components/schemas/AccountType"
        saldoInicial:
          type: number
          format: double
          minimum: 0
        estado:
          type: boolean

    AccountResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        clienteId:
          type: integer
          format: int64
        numeroCuenta:
          type: string
        tipoCuenta:
          $ref: "#/components/schemas/AccountType"
        saldoInicial:
          type: number
          format: double
        estado:
          type: boolean
      required: [id, clienteId, numeroCuenta, tipoCuenta, saldoInicial, estado]

    TransactionCreateRequest:
      type: object
      properties:
        cuentaId:
          type: integer
          format: int64
        tipoMovimiento:
          $ref: "#/components/schemas/TransactionType"
        valor:
          type: number
          format: double
          minimum: 0.01
      required: [cuentaId, tipoMovimiento, valor]

    TransactionResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        cuentaId:
          type: integer
          format: int64
        fecha:
          type: string
          format: date-time
        tipoMovimiento:
          $ref: "#/components/schemas/TransactionType"
        valor:
          type: number
          format: double
        saldo:
          type: number
          format: double
      required: [id, cuentaId, fecha, tipoMovimiento, valor, saldo]

    ReporteItemResponse:
      type: object
      properties:
        fecha:
          type: string
          format: date-time
        cliente:
          type: string
        numeroCuenta:
          type: string
        tipo:
          type: string
        saldoInicial:
          type: number
          format: double
        estado:
          type: boolean
        movimiento:
          type: number
          format: double
        saldoDisponible:
          type: number
          format: double

    ReporteResumenResponse:
      type: object
      properties:
        totalDebitos:
          type: number
          format: double
        totalCreditos:
          type: number
          format: double

    ReporteResponse:
      type: object
      properties:
        clienteId:
          type: integer
          format: int64
        fechaDesde:
          type: string
          format: date
        fechaHasta:
          type: string
          format: date
        resumen:
          $ref: "#/components/schemas/ReporteResumenResponse"
        items:
          type: array
          items:
            $ref: "#/components/schemas/ReporteItemResponse"
        pdfBase64:
          type: string
          nullable: true